test(autocomplete): recategorize and clean up browser tests
